requirements.txt
colorama == 0.4.6
python-dotenv